requirements.txt
pyTelegramBotAPI>=4.14.0
pandas>=1.5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0

